Fast PDP

Fast PDP

fastpdp(object, ...)

# S3 method for default
fastpdp(
  object,
  v,
  X,
  pred_fun = stats::predict,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

# S3 method for ranger
fastpdp(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

# S3 method for Learner
fastpdp(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

One or more variable names in X to calculate partial dependence profiles.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing \(K \ge 1\) numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional (named) arguments are passed via .... The default, stats::predict(), will work in most cases.

w

Optional vector of case weights for each row of X.

grid

Optional evaluation grid. If v is a single column name, this is a vector. Otherwise, a matrix or data.frame with length(v) columns. The default (NULL) will randomly sample rows from X[, v].

grid_type

Type of grid. If "fixed", will use unique values for discrete features, and quantile cuts otherwise.

grid_size

This is the number of rows sampled randomly from X[, v]. The result serves as evaluation grid. Only used if grid = NULL.

trim

When grid = NULL and grid_type = "fixed", numeric columns are trimmed first at those quantiles. Set to c(0, 1) to avoid trimming.

Value

An object of class "fastpdp" containing

pd: A vector, matrix or data.frame with PD values. The row order corresponds to the grid attached to the output.

grid: A vector, matrix, or data.frame representing the evaluation grid in the same row order as pd. Note that this is a sorted and deduplicated version of the input grid (if provided).

replications: How many times each value/row in the output grid was present in the input grid. Like a case weight of the result.

v: The input v.

Methods (by class)

fastpdp(default): Default Kernel SHAP method.

fastpdp(ranger): PD method for "ranger" models, see Readme for an example.

fastpdp(Learner): PD method for "mlr3" models, see Readme for an example.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)
fastpdp(fit, v = "Petal.Width", X = iris)
#> $pd
#>  [1] 4.835352 4.927042 5.018733 5.110423 5.202113 5.293803 5.660564 5.752254
#>  [9] 5.843945 5.935635 6.027325 6.119015 6.210705 6.302396 6.394086 6.485776
#> [17] 6.577466 6.669157 6.760847 6.852537 6.944227 7.035917
#> 
#> $grid
#>  [1] 0.1 0.2 0.3 0.4 0.5 0.6 1.0 1.1 1.2 1.3 1.4 1.5 1.6 1.7 1.8 1.9 2.0 2.1 2.2
#> [20] 2.3 2.4 2.5
#> 
#> $replications
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $v
#> [1] "Petal.Width"
#> 
fastpdp(fit, v = "Sepal.Width", X = iris, grid = seq(0, 2.5, by = 0.1))
#> $pd
#>  [1] 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333
#>  [9] 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333
#> [17] 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333 5.843333
#> [25] 5.843333 5.843333
#> 
#> $grid
#>  [1] 0.0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 1.0 1.1 1.2 1.3 1.4 1.5 1.6 1.7 1.8
#> [20] 1.9 2.0 2.1 2.2 2.3 2.4 2.5
#> 
#> $replications
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $v
#> [1] "Sepal.Width"
#> 

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
fastpdp(fit, v = "Petal.Width", X = iris)
#> $pd
#>       Sepal.Length Sepal.Width
#>  [1,]     5.849924    2.371959
#>  [2,]     5.849325    2.434303
#>  [3,]     5.848725    2.496648
#>  [4,]     5.848126    2.558992
#>  [5,]     5.847526    2.621337
#>  [6,]     5.846927    2.683682
#>  [7,]     5.844528    2.933060
#>  [8,]     5.843929    2.995404
#>  [9,]     5.843329    3.057749
#> [10,]     5.842730    3.120094
#> [11,]     5.842130    3.182438
#> [12,]     5.841531    3.244783
#> [13,]     5.840931    3.307127
#> [14,]     5.840332    3.369472
#> [15,]     5.839732    3.431816
#> [16,]     5.839133    3.494161
#> [17,]     5.838533    3.556505
#> [18,]     5.837933    3.618850
#> [19,]     5.837334    3.681195
#> [20,]     5.836734    3.743539
#> [21,]     5.836135    3.805884
#> [22,]     5.835535    3.868228
#> 
#> $grid
#>  [1] 0.1 0.2 0.3 0.4 0.5 0.6 1.0 1.1 1.2 1.3 1.4 1.5 1.6 1.7 1.8 1.9 2.0 2.1 2.2
#> [20] 2.3 2.4 2.5
#> 
#> $replications
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $v
#> [1] "Petal.Width"
#> 
fastpdp(fit, v = c("Petal.Width", "Species"), X = iris, grid_type = "random")
#> $pd
#>       Sepal.Length Sepal.Width
#>  [1,]     7.086927    3.692112
#>  [2,]     7.086327    3.754457
#>  [3,]     7.085728    3.816801
#>  [4,]     7.085128    3.879146
#>  [5,]     5.483170    2.489110
#>  [6,]     5.481971    2.613799
#>  [7,]     5.481371    2.676143
#>  [8,]     5.480772    2.738488
#>  [9,]     5.480172    2.800832
#> [10,]     5.479572    2.863177
#> [11,]     4.964088    2.555613
#> [12,]     4.963489    2.617958
#> [13,]     4.962889    2.680302
#> [14,]     4.962289    2.742647
#> [15,]     4.961690    2.804991
#> [16,]     4.961090    2.867336
#> [17,]     4.960491    2.929680
#> 
#> $grid
#>     Petal.Width    Species
#> 38          0.1     setosa
#> 4           0.2     setosa
#> 41          0.3     setosa
#> 27          0.4     setosa
#> 94          1.0 versicolor
#> 74          1.2 versicolor
#> 98          1.3 versicolor
#> 64          1.4 versicolor
#> 79          1.5 versicolor
#> 84          1.6 versicolor
#> 139         1.8  virginica
#> 143         1.9  virginica
#> 148         2.0  virginica
#> 140         2.1  virginica
#> 118         2.2  virginica
#> 116         2.3  virginica
#> 115         2.4  virginica
#> 
#> $replications
#>  [1] 1 6 3 1 1 2 4 3 3 1 2 1 2 1 2 2 1
#> 
#> $v
#> [1] "Petal.Width" "Species"    
#>