Friedman's H

Friedman's H

fx_friedmans_h(object, ...)

# S3 method for default
fx_friedmans_h(
  object,
  v,
  X,
  pred_fun = stats::predict,
  normalize = TRUE,
  take_sqrt = TRUE,
  grid_size = 200L,
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for ranger
fx_friedmans_h(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  normalize = TRUE,
  take_sqrt = TRUE,
  grid_size = 200L,
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for Learner
fx_friedmans_h(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  normalize = TRUE,
  take_sqrt = TRUE,
  grid_size = 200L,
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  verbose = TRUE,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

Vector of variables names to calculate Friedman's H for.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing \(K \ge 1\) numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional (named) arguments are passed via .... The default, stats::predict(), will work in most cases.

normalize

TODO

take_sqrt

TODO

grid_size

Determines the grid When grid = NULL. Character/factor variables are evaluated at each unique value. A numeric v with more than grid_size unique values is evaluated at grid_size quantiles. If v has length \(p > 1\), the \(p\)th root of grid_size is used instead.

n_max

If X has more rows than n_max, a random sample of n_max rows is selected.

out_names

Names of the output columns.

w

Optional vector of case weights for each row of X.

verbose

Should a progress bar be shown? Default is TRUE.

Value

A data.frame

Methods (by class)

fx_friedmans_h(default): Default method.

fx_friedmans_h(ranger): Method for "ranger" models, see Readme for an example.

fx_friedmans_h(Learner): Method for "mlr3" models, see Readme for an example.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
fx_friedmans_h(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
#>              V      Stat
#> 1  Sepal.Width 0.0000000
#> 2 Petal.Length 0.0000000
#> 3  Petal.Width 0.2241348
#> 4      Species 0.2241348

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
h <- fx_friedmans_h(fit, v = colnames(iris[3:5]), X = iris)
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%
h
#>              V Sepal.Length Sepal.Width
#> 1 Petal.Length    0.0000000   0.0000000
#> 2  Petal.Width    0.2181502   0.1990873
#> 3      Species    0.2181502   0.1990873
fx_friedmans_h(fit, v = c("Petal.Width", "Species"), X = iris)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%
#>             V Sepal.Length Sepal.Width
#> 1 Petal.Width    0.2181502   0.1990873
#> 2     Species    0.2181502   0.1990873