PD Importance

The sample variance of the partial dependence function of a feature \(X^{(j)}\) evaluated over a data set \(D\) can be used as measure of (main effect) importance, defined as \(PDI(j, D) = \textrm{Var}\left(\sum_{i \in D} \textrm{PD}_{j}(x_i^{(j)})\right)\). Similarly, we can define the PD importance of two or more features together, measuring their joint effect (main effect plus interaction).

fx_importance(object, ...)

# S3 method for default
fx_importance(
  object,
  v,
  X,
  pred_fun = stats::predict,
  n_max = 300L,
  out_names = NULL,
  w = NULL,
  sort = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for ranger
fx_importance(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  n_max = 300L,
  out_names = NULL,
  w = NULL,
  sort = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for Learner
fx_importance(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  n_max = 300L,
  out_names = NULL,
  w = NULL,
  sort = TRUE,
  verbose = TRUE,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

Vector or list of feature names for which PD importance is to be calculated. If passed as list, vectors of feature names are evaluted together.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing \(K \ge 1\) numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional (named) arguments are passed via .... The default, stats::predict(), will work in most cases.

n_max

If X has more than n_max rows, a random sample of n_max rows is selected for calculations.

out_names

Names of the output columns.

w

Optional vector of case weights for each row of X.

sort

Should the result be sortedby importance? Default is TRUE.

verbose

Should a progress bar be shown? Default is TRUE.

Value

A data.frame with one column holding the feature name(s), and K columns with the PD importance for the K components of the predict function.

Methods (by class)

fx_importance(default): Default method.

fx_importance(ranger): Method for "ranger" models .

fx_importance(Learner): Method for "mlr3" models.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
fx_importance(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
#>              V        Imp
#> 2 Petal.Length 2.14289434
#> 4      Species 0.18579991
#> 3  Petal.Width 0.05770716
#> 1  Sepal.Width 0.04671705

v <- list("Sepal.Width", c("Petal.Width", "Petal.Length"), "Species")
fx_importance(fit, v = v, X = iris, verbose = FALSE)
#>                          V        Imp
#> 2 Petal.Width:Petal.Length 1.52341106
#> 3                  Species 0.18579991
#> 1              Sepal.Width 0.04671705

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
fx_importance(fit, v = v, X = iris)
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#>              V Sepal.Length Sepal.Width
#> 1 Petal.Length 2.557647e+00  0.07455559
#> 3      Species 8.146001e-01  0.90860036
#> 2  Petal.Width 2.088418e-05  0.22582807

# MODEL THREE: Gamma GLM with log link
fit <- glm(Sepal.Length ~ ., data = iris, family = Gamma(link = log))

fx_importance(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
#>              V         Imp
#> 2 Petal.Length 0.051261578
#> 4      Species 0.002540002
#> 1  Sepal.Width 0.001674607
#> 3  Petal.Width 0.001402744

# On original scale
fx_importance(fit, v = names(iris[-1]), X = iris, verbose = FALSE, type = "response")
#>              V        Imp
#> 2 Petal.Length 1.69518081
#> 4      Species 0.08922822
#> 1  Sepal.Width 0.05823951
#> 3  Petal.Width 0.04861255