Fast Partial Dependence Function

Fast implementation of Friedman's (empirical) partial dependence (PD) function of feature subset \(J\), given by $$\textrm{PD}_J(v) = \frac{1}{|D|} \sum_{i \in D} \hat f(v, x_{i,\setminus J})$$, where

\(D\) is a reference data set,

\(J\) is a index set

\(\hat f\) is the fitted model, and

\(x_{i,\setminus J}\) is the feature vector of the i-th observation without components in \(J\) (which are replaced by the function argument(s) \(v\)).

The function supports both

multivariate predictions (e.g., multi-classification settings) and

multivariate grids.

fx_pd(object, ...)

# S3 method for default
fx_pd(
  object,
  v,
  X,
  pred_fun = stats::predict,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for ranger
fx_pd(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for Learner
fx_pd(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  verbose = TRUE,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

Vector of feature names.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

grid

Named list. Each element specifies the evaluation grid for the corresponding feature. Missing components are automatically added. If v has length 1, then grid can also be a vector.

grid_size

Controls the grid size for variables not in grid.

trim

Variables not in grid are trimmed at those two quantiles. Set to c(0, 1) for no trimming.

strategy

How should evaluation points of variables not in grid be found? Either "quantile" or "uniform".

n_max

If X has more than n_max rows, a random sample of n_max rows is selected for calculations (after determining grid).

w

Optional vector of case weights for each row of X.

verbose

Should a progress bar be shown? The default is TRUE.

Value

An object of class "fx_pd", containing these elements:

grid: Named list of evaluation points.

pd: Named list of PD matrices.

v: Same as input v.

Methods (by class)

fx_pd(default): Default method.

fx_pd(ranger): Method for "ranger" models.

fx_pd(Learner): Method for "mlr3" models.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
pd <- fx_pd(fit, v = names(iris[-1]), X = iris)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%
pd
#> PD values. Use summary(pd, variable) to extract the results.
head(summary(pd))
#>   Sepal.Width     pred
#> 1         2.0 5.319013
#> 2         2.2 5.418191
#> 3         2.3 5.467780
#> 4         2.4 5.517369
#> 5         2.5 5.566958
#> 6         2.6 5.616547
summary(pd, "Species")
#>      Species     pred
#> 1     setosa 6.425687
#> 2 versicolor 5.702125
#> 3  virginica 5.402189

pd <- fx_pd(fit, v = "Petal.Width", X = iris, grid = seq(0, 1, by = 0.5))
summary(pd)
#>   Petal.Width     pred
#> 1         0.0 6.221309
#> 2         0.5 6.063732
#> 3         1.0 5.906154
summary(fx_pd(fit, v = "Petal.Width", X = iris, grid = seq(1, 0, by = -0.5)))
#>   Petal.Width     pred
#> 1         1.0 5.906154
#> 2         0.5 6.063732
#> 3         0.0 6.221309
summary(fx_pd(fit, v = "Species", X = iris))
#>      Species     pred
#> 1     setosa 6.425687
#> 2 versicolor 5.702125
#> 3  virginica 5.402189

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
v <- names(iris[3:5])
partial_grid <- list(Petal.Width = seq(0, 1, by = 0.5))
pd <- fx_pd(fit, v = v, X = iris, grid = partial_grid, verbose = FALSE)
summary(pd, "Species")
#>      Species Sepal.Length Sepal.Width
#> 1     setosa     7.080336    4.377487
#> 2 versicolor     5.481975    2.613383
#> 3  virginica     4.967689    2.181130
summary(pd, "Petal.Width")
#>   Petal.Width Sepal.Length Sepal.Width
#> 1         0.0     5.850524    2.309614
#> 2         0.5     5.847526    2.621337
#> 3         1.0     5.844528    2.933060
head(summary(pd, "Petal.Length"))
#>   Petal.Length Sepal.Length Sepal.Width
#> 1          1.1     3.435329    2.646205
#> 2          1.3     3.616518    2.677141
#> 3          1.4     3.707113    2.692608
#> 4          1.5     3.797708    2.708076
#> 5          1.6     3.888302    2.723543
#> 6          1.7     3.978897    2.739011

# MODEL THREE: Gamma GLM with log link
fit <- glm(
  Sepal.Length ~ . + Petal.Width:Species, 
  data = iris, 
  family = Gamma(link = log)
)
summary(fx_pd(fit, v = "Species", X = iris, type = "response"))
#>      Species     pred
#> 1     setosa 6.989680
#> 2 versicolor 5.828742
#> 3  virginica 5.494072