Fast PDP

Fast PDP

fx_pdp(object, ...)

# S3 method for default
fx_pdp(
  object,
  v,
  X,
  pred_fun = stats::predict,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)

# S3 method for ranger
fx_pdp(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)

# S3 method for Learner
fx_pdp(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  n_max = 500L,
  out_names = NULL,
  w = NULL,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

One or more variable names in X to calculate partial dependence profiles.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing \(K \ge 1\) numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional (named) arguments are passed via .... The default, stats::predict(), will work in most cases.

grid

Optional evaluation grid. If v is a single column name, this is a vector/factor. Otherwise, a matrix or data.frame with length(v) columns.

grid_size

Determines the grid When grid = NULL. Character/factor variables are evaluated at each unique value. A numeric v with more than grid_size unique values is evaluated at grid_size quantiles. If v has length \(p > 1\), the \(p\)th root of grid_size is used instead.

trim

A vector with two probabilities used to trim non-discrete numeric v before applying quantile binning (only if grid = NULL). Set to c(0, 1) to avoid trimming.

n_max

If X has more rows than n_max, a random sample of n_max rows is selected.

out_names

Names of the output columns.

w

Optional vector of case weights for each row of X.

Value

A dataframe with partial dependence per grid value.

Methods (by class)

fx_pdp(default): Default method.

fx_pdp(ranger): Method for "ranger" models, see Readme for an example.

fx_pdp(Learner): Method for "mlr3" models, see Readme for an example.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)
pd <- fx_pdp(fit, v = "Petal.Width", X = iris)
pd[1:4, ]
#>   Petal.Width     pred
#> 1         0.1 4.835352
#> 2         0.2 4.927042
#> 3         0.3 5.018733
#> 4         0.4 5.110423

fx_pdp(fit, v = "Petal.Width", X = iris, grid = seq(0, 1, by = 0.5), out_name = "P")
#>   Petal.Width        P
#> 1         0.0 4.743662
#> 2         0.5 5.202113
#> 3         1.0 5.660564
fx_pdp(fit, v = "Petal.Width", X = iris, grid = seq(1, 0, by = -0.5))
#>   Petal.Width     pred
#> 1         1.0 5.660564
#> 2         0.5 5.202113
#> 3         0.0 4.743662
fx_pdp(fit, v = "Species", X = iris)
#>      Species     pred
#> 1     setosa 5.880113
#> 2 versicolor 5.819859
#> 3  virginica 5.830028

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
fx_pdp(fit, v = "Species", X = iris)
#>      Species Sepal.Length Sepal.Width
#> 1     setosa     7.080336    4.377487
#> 2 versicolor     5.481975    2.613383
#> 3  virginica     4.967689    2.181130
pd <- fx_pdp(fit, v = c("Petal.Width", "Species"), X = iris)
pd[1:4, ]
#>   Petal.Width Species Sepal.Length Sepal.Width
#> 1         0.1  setosa     7.086927    3.692112
#> 2         0.2  setosa     7.086327    3.754457
#> 3         1.2  setosa     7.080332    4.377902
#> 4         1.5  setosa     7.078533    4.564936