Fast Friedman's H

Fast Friedman's H

pd_interaction(object, ...)

# S3 method for default
pd_interaction(
  object,
  v,
  X,
  pred_fun = stats::predict,
  pairwise = FALSE,
  n_max = 200L,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for ranger
pd_interaction(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  pairwise = FALSE,
  n_max = 200L,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for Learner
pd_interaction(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  pairwise = FALSE,
  n_max = 200L,
  w = NULL,
  verbose = TRUE,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

Vector of feature names for which interaction statistics are to be calculated.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

pairwise

The default (FALSE) calculates overall interaction strength per feature. Set to TRUE to get pairwise statistics (slower).

n_max

If X has more than n_max rows, a random sample of n_max rows is selected for calculations (after determining grid).

w

Optional vector of case weights for each row of X.

verbose

Should a progress bar be shown? The default is TRUE.

Value

An object of class "pd_interaction", containing these elements:

num: Matrix with squared numerator values.

denom: Matrix with squared denominator values.

v: Same as input v.

pairwise: Same as input pairwise.

Methods (by class)

pd_interaction(default): Default method.

pd_interaction(ranger): Method for "ranger" models .

pd_interaction(Learner): Method for "mlr3" models.

References

Friedman, J. H. and Popescu, B. E. (2008). "Predictive learning via rule ensembles." The Annals of Applied Statistics. JSTOR, 916–54.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- pd_interaction(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
inter
#> Interaction statistics. Use summary(...) to extract the results.
summary(inter)
#> Normalized Friedman's H (sorted) 
#>                   Stat
#> Species      0.2241348
#> Petal.Width  0.2241348
#> Sepal.Width  0.0000000
#> Petal.Length 0.0000000
summary(inter, normalize = FALSE, squared = TRUE)
#> Unnormalized Friedman's H squared (sorted) 
#>                    Stat
#> Species      0.02980153
#> Petal.Width  0.02980153
#> Sepal.Width  0.00000000
#> Petal.Length 0.00000000

inter <- pd_interaction(
  fit, v = names(iris[-1]), X = iris, verbose = FALSE, pairwise = TRUE
)
summary(inter)
#> Normalized Friedman's H (sorted) 
#>                               Stat
#> Petal.Width:Species      0.2355031
#> Sepal.Width:Petal.Length 0.0000000
#> Sepal.Width:Petal.Width  0.0000000
#> Sepal.Width:Species      0.0000000
#> Petal.Length:Petal.Width 0.0000000
#> Petal.Length:Species     0.0000000

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
pd_interaction(fit, v = v, X = iris)
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%
#> Interaction statistics. Use summary(...) to extract the results.
pd_interaction(fit, v = v, X = iris, pairwise = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%
#> Interaction statistics. Use summary(...) to extract the results.

# MODEL THREE: Gamma GLM with log link
fit <- glm(
  Sepal.Length ~ . + Petal.Width:Species, 
  data = iris, 
  family = Gamma(link = log)
)

# No interactions for additive features, at least on link scale
pd_interaction(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
#> Interaction statistics. Use summary(...) to extract the results.

# On original scale, we have interactions everywhere...
pd_interaction(
  fit, v = names(iris[-1]), X = iris, verbose = FALSE, type = "response"
)
#> Interaction statistics. Use summary(...) to extract the results.