Partial Dependence Profiles

For a set v of features, their individual partial dependence is evaluated over evaluation grids. The grids can be prespecified for any subset of the features. The function calls the prediction function m times, where m is the number of features. Multivariate predictions are supported.

pd_profiles(object, ...)

# S3 method for default
pd_profiles(
  object,
  v,
  X,
  pred_fun = stats::predict,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

# S3 method for ranger
pd_profiles(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

# S3 method for Learner
pd_profiles(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...), for instance type = "response" in a glm() model.

v

Vector of feature names.

X

A data.frame or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

grid

Named list. Each element specifies the evaluation grid for the corresponding feature. Missing components are automatically generated via univariate_grid(). If v has length 1, then grid can also be a vector.

grid_size

Approximate grid size.

trim

A non-discrete numeric z is trimmed at these quantile probabilities before calculations. Set to c(0, 1) for no trimming.

strategy

How to find evaluation points of non-discrete numeric columns? Either "quantile" or "uniform" (via pretty()), see description of univariate_grid().

n_max

If X has more than n_max rows, a random sample of n_max rows is selected from X. In this case, set a random seed for reproducibility.

w

Optional vector of case weights for each row of X.

Value

A list of PD profiles per variable in v. Has additional class "pd_profiles".

Methods (by class)

pd_profiles(default): Default method.

pd_profiles(ranger): Method for "ranger" models.

pd_profiles(Learner): Method for "mlr3" models.

References

Friedman, Jerome H. "Greedy Function Approximation: A Gradient Boosting Machine." Annals of Statistics 29 (2000): 1189-1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
pd <- pd_profiles(fit, v = names(iris[-1]), X = iris)
pd$Species
#>      Species        y
#> 1     setosa 6.425687
#> 2 versicolor 5.702125
#> 3  virginica 5.402189
head(pd$Sepal.Width)
#>   Sepal.Width        y
#> 1         2.0 5.319013
#> 2         2.2 5.418191
#> 3         2.3 5.467780
#> 4         2.4 5.517369
#> 5         2.5 5.566958
#> 6         2.6 5.616547

pd <- pd_profiles(fit, v = "Petal.Width", X = iris, grid = seq(1, 0, by = -0.5))
pd$Petal.Width 
#>   Petal.Width        y
#> 1         1.0 5.906154
#> 2         0.5 6.063732
#> 3         0.0 6.221309

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
v <- names(iris[3:5])
partial_grid <- list(Petal.Width = seq(0, 1, by = 0.5))
pd <- pd_profiles(fit, v = v, X = iris, grid = partial_grid, verbose = FALSE)
pd$Species
#>      Species Sepal.Length Sepal.Width
#> 1     setosa     7.080336    4.377487
#> 2 versicolor     5.481975    2.613383
#> 3  virginica     4.967689    2.181130
 
# MODEL THREE: Gamma GLM -> pass options to predict() via ...
fit <- glm(
  Sepal.Length ~ . + Petal.Width:Species, 
  data = iris, 
  family = Gamma(link = log)
)
pd_profiles(fit, v = "Species", X = iris, type = "response")$Species
#>      Species        y
#> 1     setosa 6.989680
#> 2 versicolor 5.828742
#> 3  virginica 5.494072