Barebone Partial Dependence Function

Creates partial dependence values for a given model, data, and grid. This is the workhorse function behind pd(), fx_interaction(), and fx_importance().

It is fast because:

There is a single call to pred_fun().

If more than 5% of grid rows are duplicated, they are dropped from the calculation. Resulting PD values are mapped back. This is important for fx_interaction() and fx_importance() that use sampled data as grid.

If more than 5% of non-grid rows are duplicated, these are dropped and compensated by summing up their case weights.

pd_raw(
  object,
  v,
  X,
  pred_fun,
  grid,
  w = NULL,
  compress_X = TRUE,
  compress_grid = TRUE,
  ...
)

Arguments

object

Fitted model object.

v

Vector of variable name(s) to calculate partial dependence for.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

grid

A vector (if length(v) == 1L), or a matrix/data.frame otherwise.

w

Optional vector of case weights for each row of X.

compress_X

If X has a single non-v column: should duplicates be removed and compensated via case weights? (Applied only if >5% duplicates.) Default is TRUE.

compress_grid

Should duplicates in grid be removed, and resulting PDs mapped back to the original grid index? (Applied only if >5% duplicates.) Default is TRUE.

...

Additional arguments passed to pred_fun(object, X, ...).

Value

A matrix of partial dependence values (one column per prediction dimension, one row per grid row).

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
pdp_raw(fit, v = "Petal.Width", X = iris, pred_fun = predict, grid = 1:2)
#> Error in pdp_raw(fit, v = "Petal.Width", X = iris, pred_fun = predict,     grid = 1:2): could not find function "pdp_raw"