Fast PDP

Fast PDP

fastpdp(object, ...)

# S3 method for default
fastpdp(
  object,
  v,
  X,
  pred_fun = stats::predict,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

# S3 method for ranger
fastpdp(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

# S3 method for Learner
fastpdp(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  w = NULL,
  grid = NULL,
  grid_type = c("fixed", "random"),
  grid_size = 27L,
  trim = c(0.01, 0.99),
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

One or more variable names in X to calculate partial dependence profiles.

X

Dataframe or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing \(K \ge 1\) numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional (named) arguments are passed via .... The default, stats::predict(), will work in most cases.

w

Optional vector of case weights for each row of X.

grid

Optional evaluation grid. If v is a single column name, this is a vector. Otherwise, a matrix or data.frame with length(v) columns. The default (NULL) will randomly sample rows from X[, v].

grid_type

Type of grid. If "fixed", will use unique values for discrete features, and quantile cuts otherwise.

grid_size

This is the number of rows sampled randomly from X[, v]. The result serves as evaluation grid. Only used if grid = NULL.

trim

When grid = NULL and grid_type = "fixed", numeric columns are trimmed first at those quantiles. Set to c(0, 1) to avoid trimming.

Value

An object of class "fastpdp" containing

pd: A vector, matrix or data.frame with PD values. The row order corresponds to the grid attached to the output.

grid: A vector, matrix, or data.frame representing the evaluation grid in the same row order as pd. Note that this is a sorted and deduplicated version of the input grid (if provided).

replications: How many times each value/row in the output grid was present in the input grid. Like a case weight of the result.

v: The input v.

Methods (by class)

fastpdp(default): Default Kernel SHAP method.

fastpdp(ranger): PD method for "ranger" models, see Readme for an example.

fastpdp(Learner): PD method for "mlr3" models, see Readme for an example.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ Species + Petal.Width, data = iris)
fastpdp(fit, v = "Petal.Width", X = iris)
#> Partial dependence profile of dimension 22 x 1
fastpdp(fit, v = "Sepal.Width", X = iris, grid = seq(0, 2.5, by = 0.1))
#> Partial dependence profile of dimension 26 x 1

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
fastpdp(fit, v = "Petal.Width", X = iris)
#> Partial dependence profile of dimension 22 x 2
fastpdp(fit, v = c("Petal.Width", "Species"), X = iris, grid_type = "random")
#> Partial dependence profile of dimension 15 x 2