Fast Partial Dependence Function

Fast implementation of Friedman's (empirical) partial dependence (PD) function of feature subset \(J\), given by $$\textrm{PD}_J(v) = \frac{1}{|D|} \sum_{i \in D} \hat f(v, x_{i,\setminus J})$$, where

\(D\) is a reference data set,

\(J\) is a index set

\(\hat f\) is the fitted model, and

\(x_{i,\setminus J}\) is the feature vector of the i-th observation without components in \(J\) (which are replaced by the function argument(s) \(v\)).

The function supports both

multivariate predictions (e.g., multi-classification settings) and

multivariate grids.

pd_profiles(object, ...)

# S3 method for default
pd_profiles(
  object,
  v,
  X,
  pred_fun = stats::predict,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for ranger
pd_profiles(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  verbose = TRUE,
  ...
)

# S3 method for Learner
pd_profiles(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  grid = NULL,
  grid_size = 36L,
  trim = c(0.01, 0.99),
  strategy = c("quantile", "uniform"),
  n_max = 1000L,
  w = NULL,
  verbose = TRUE,
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...).

v

Vector of feature names.

X

A data.frame or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

grid

Named list. Each element specifies the evaluation grid for the corresponding feature. Missing components are automatically added. If v has length 1, then grid can also be a vector.

grid_size

Controls the approximate grid size. If X has p columns, then each (non-discrete) column will be reduced to about the p-th root of grid_size values.

trim

Non-discrete columns are trimmed at corresponding quantiles. Set to c(0, 1) for no trimming.

strategy

How should evaluation points of non-discrete columns be found? Either "quantile" or "uniform".

n_max

If X has more than n_max rows, a random sample of n_max rows is selected for calculations (after determining grid).

w

Optional vector of case weights for each row of X.

verbose

Should a progress bar be shown? The default is TRUE.

Value

An object of class "pd_profiles", containing these elements:

grid: Named list of evaluation points.

pd: Named list of PD matrices.

v: Same as input v.

Methods (by class)

pd_profiles(default): Default method.

pd_profiles(ranger): Method for "ranger" models.

pd_profiles(Learner): Method for "mlr3" models.

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
pd <- pd_profiles(fit, v = names(iris[-1]), X = iris)
#> Error in .basic_check(X = X, v = v, pred_fun = pred_fun, w = w): could not find function ".basic_check"
pd
#> Error in eval(expr, envir, enclos): object 'pd' not found
head(summary(pd))
#> Error in eval(expr, envir, enclos): object 'pd' not found
summary(pd, "Species")
#> Error in eval(expr, envir, enclos): object 'pd' not found

pd <- pd_profiles(fit, v = "Petal.Width", X = iris, grid = seq(0, 1, by = 0.5))
#> Error in .basic_check(X = X, v = v, pred_fun = pred_fun, w = w): could not find function ".basic_check"
summary(pd)
#> Error in eval(expr, envir, enclos): object 'pd' not found
summary(pd_profiles(fit, v = "Petal.Width", X = iris, grid = seq(1, 0, by = -0.5)))
#> Error in .basic_check(X = X, v = v, pred_fun = pred_fun, w = w): could not find function ".basic_check"
summary(pd_profiles(fit, v = "Species", X = iris))
#> Error in .basic_check(X = X, v = v, pred_fun = pred_fun, w = w): could not find function ".basic_check"

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
v <- names(iris[3:5])
partial_grid <- list(Petal.Width = seq(0, 1, by = 0.5))
pd <- pd_profiles(fit, v = v, X = iris, grid = partial_grid, verbose = FALSE)
#> Error in .basic_check(X = X, v = v, pred_fun = pred_fun, w = w): could not find function ".basic_check"
summary(pd, "Species")
#> Error in eval(expr, envir, enclos): object 'pd' not found
summary(pd, "Petal.Width")
#> Error in eval(expr, envir, enclos): object 'pd' not found
head(summary(pd, "Petal.Length"))
#> Error in eval(expr, envir, enclos): object 'pd' not found

# MODEL THREE: Gamma GLM with log link
fit <- glm(
  Sepal.Length ~ . + Petal.Width:Species, 
  data = iris, 
  family = Gamma(link = log)
)
summary(pd_profiles(fit, v = "Species", X = iris, type = "response"))
#> Error in .basic_check(X = X, v = v, pred_fun = pred_fun, w = w): could not find function ".basic_check"