Barebone Partial Dependence (PD) Function

Used internally by pd_profiles(), pd_interaction(), and pd_importance(). Furthermore, it can be used to calculate multi-dimensional partial dependencies. In this case, provide your own grid, for instance created via make_grid().

pd_raw(
  object,
  v,
  X,
  grid,
  pred_fun = stats::predict,
  n_max = 1000L,
  w = NULL,
  compress_X = TRUE,
  compress_grid = TRUE,
  check = TRUE,
  ...
)

Arguments

object

Fitted model object.

v

Vector of feature names.

X

A data.frame or matrix serving as background dataset.

grid

A vector (if length(v) == 1L), or a matrix/data.frame otherwise.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

n_max

If X has more than n_max rows, a random sample of n_max rows is selected for calculations (after determining grid).

w

Optional vector of case weights for each row of X.

compress_X

If X has a single non-v column: should duplicates be removed and compensated via case weights? Applied if >5% duplicates. Default is TRUE.

compress_grid

Should duplicates in grid be removed and PDs mapped back to the original grid index? Applied if >5% duplicates. Default is TRUE.

check

Should input checks be applied? Default is TRUE.

...

Additional arguments passed to pred_fun(object, X, ...).

Value

A matrix of partial dependence values (one column per prediction dimension, one row per grid row).

References

Friedman J. H. (2001). Greedy function approximation: A gradient boosting machine. The Annals of Statistics, 29:1189–1232.

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
pd_raw(fit, v = "Petal.Width", X = iris, pred_fun = predict, grid = 1:2)
#>          [,1]
#> [1,] 5.906154
#> [2,] 5.590999

# 2D PD
grid <- make_grid(iris[4:5])
#> Error in make_grid(iris[4:5]): could not find function "make_grid"
grid
#> function (nx = NULL, ny = nx, col = "lightgray", lty = "dotted", 
#>     lwd = par("lwd"), equilogs = TRUE) 
#> {
#>     if (is.null(nx) || (!is.na(nx) && nx >= 1)) {
#>         log <- par("xlog")
#>         if (is.null(nx)) {
#>             ax <- par("xaxp")
#>             if (log && equilogs && ax[3L] > 0) 
#>                 ax[3L] <- 1
#>             at <- axTicks(1, axp = ax, log = log)
#>         }
#>         else {
#>             U <- par("usr")
#>             at <- seq.int(U[1L], U[2L], length.out = nx + 1)
#>             at <- (if (log) 
#>                 10^at
#>             else at)[-c(1, nx + 1)]
#>         }
#>         abline(v = at, col = col, lty = lty, lwd = lwd)
#>     }
#>     if (is.null(ny) || (!is.na(ny) && ny >= 1)) {
#>         log <- par("ylog")
#>         if (is.null(ny)) {
#>             ax <- par("yaxp")
#>             if (log && equilogs && ax[3L] > 0) 
#>                 ax[3L] <- 1
#>             at <- axTicks(2, axp = ax, log = log)
#>         }
#>         else {
#>             U <- par("usr")
#>             at <- seq.int(U[3L], U[4L], length.out = ny + 1)
#>             at <- (if (log) 
#>                 10^at
#>             else at)[-c(1, ny + 1)]
#>         }
#>         abline(h = at, col = col, lty = lty, lwd = lwd)
#>     }
#> }
#> <bytecode: 0x559942bc5118>
#> <environment: namespace:graphics>
values <- pd_raw(fit, X = iris, v = colnames(iris)[4:5], grid = grid)
#> Error in .check_grid(g = grid, v = v, X_is_matrix = is.matrix(X)): NCOL(grid) must equal length(v)
result <- data.frame(grid, values)
#> Error in eval(expr, envir, enclos): object 'values' not found
head(result)
#> Error in eval(expr, envir, enclos): object 'result' not found