Barebone Partial Dependence (PD) Function

Workhorse of pd_profiles() and interaction_statistics(). Furthermore, it can be used to calculate multi-dimensional partial dependencies. In this case, provide your own evaluation grid, for instance via multivariate_grid().

pd_raw(
  object,
  v,
  X,
  grid,
  pred_fun = stats::predict,
  n_max = 1000L,
  w = NULL,
  compress_X = TRUE,
  compress_grid = TRUE,
  check = TRUE,
  ...
)

Arguments

object

Fitted model object.

v

Vector of feature names.

X

A data.frame or matrix serving as background dataset.

grid

A vector (if length(v) == 1L), or a matrix/data.frame otherwise.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

n_max

If X has more than n_max rows, a random sample of n_max rows is selected from X. In this case, set a random seed for reproducibility.

w

Optional vector of case weights for each row of X.

compress_X

If X has a single non-v column: should duplicates be removed and compensated via case weights? Default is TRUE.

compress_grid

Should duplicates in grid be removed and PDs mapped back to the original grid index? Default is TRUE.

check

Should input checks be applied? Default is TRUE.

...

Additional arguments passed to pred_fun(object, X, ...), for instance type = "response" in a glm() model.

Value

A matrix of partial dependence values (one column per prediction dimension, one row per grid row, in the same order as grid).

References

Friedman, Jerome H. "Greedy Function Approximation: A Gradient Boosting Machine." Annals of Statistics 29 (2000): 1189-1232.

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
pd_raw(fit, v = "Petal.Width", X = iris, pred_fun = predict, grid = 1:2)
#>             y
#> [1,] 5.906154
#> [2,] 5.590999

# 2D PD
grid <- multivariate_grid(iris[4:5])
grid
#>    Petal.Width    Species
#> 1          0.1     setosa
#> 2          0.2     setosa
#> 3          1.2     setosa
#> 4          1.5     setosa
#> 5          1.9     setosa
#> 6          2.5     setosa
#> 7          0.1 versicolor
#> 8          0.2 versicolor
#> 9          1.2 versicolor
#> 10         1.5 versicolor
#> 11         1.9 versicolor
#> 12         2.5 versicolor
#> 13         0.1  virginica
#> 14         0.2  virginica
#> 15         1.2  virginica
#> 16         1.5  virginica
#> 17         1.9  virginica
#> 18         2.5  virginica
values <- pd_raw(fit, X = iris, v = colnames(iris)[4:5], grid = grid)
result <- data.frame(grid, values)
head(result, 2)
#>   Petal.Width Species        y
#> 1         0.1  setosa 6.772147
#> 2         0.2  setosa 6.740632